test/tempfile.cpp
