+++ test/meson.build
@@ -16,7 +16,7 @@ tests = [
 
 if gtest_dep.found() and not meson.is_cross_build()
     foreach test_name : tests
-        test_exe = executable(test_name, [test_name+'.cpp'],
+        test_exe = executable(test_name, [test_name+'.cpp', 'tempfile.cpp'],
                               implicit_include_directories: false,
                               include_directories : [include_directory, src_directory],
                               link_with : libzim,
